check http and ssh fix and cleanup (for vpc pfrules tests)
