+++ tools/marvin/marvin/cloudstackTestCase.py
@@ -14,11 +14,11 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-
+import time
 import unittest
 from marvin.lib.utils import verifyElementInList
 from marvin.codes import PASS
-from marvin.lib.base import VirtualMachine
+from marvin.lib.base import (VirtualMachine, NATRule, NetworkACL)
 
 
 def user(Name, DomainName, AcctType):
@@ -45,7 +45,7 @@ class cloudstackTestCase(unittest.case.TestCase):
         @desc:Uses the utility function verifyElementInList and
         asserts based upon PASS\FAIL value of the output.
         Takes one additional argument of what message to assert with
-        when failed
+        when faileddef start_vpcrouter(
         '''
         out = verifyElementInList(inp, toverify, responsevar, pos)
         unittest.TestCase.assertEqual(out[0], PASS, "msg:%s" % out[1])
@@ -97,3 +97,99 @@ class cloudstackTestCase(unittest.case.TestCase):
                 obj.delete(api_client, expunge=True)
             else:
                 obj.delete(api_client)
+
+    def check_wget_from_vm(self, vm, public_ip, network=None, testnegative=False, isVmAccessible=True):
+        import urllib.request, urllib.error
+        self.debug(f"Checking if we can wget from a VM={vm.name} http server on public_ip={public_ip.ipaddress.ipaddress}, expecting failure == {testnegative} and vm is acceccible == {isVmAccessible}")
+        try:
+            if not isVmAccessible:
+                self.create_natrule_for_services(vm, public_ip, network)
+            self.setup_webserver(vm)
+
+            urllib.request.urlretrieve(f"http://{public_ip.ipaddress.ipaddress}/test.html", filename="test.html")
+            if not testnegative:
+                self.debug("Successesfull to wget from VM=%s http server on public_ip=%s" % (vm.name, public_ip.ipaddress.ipaddress))
+            else:
+                self.fail("Successesfull to wget from VM=%s http server on public_ip=%s" % (vm.name, public_ip.ipaddress.ipaddress))
+        except Exception as e:
+            if not testnegative:
+                self.fail("Failed to wget from VM=%s http server on public_ip=%s: %s" % (vm.name, public_ip.ipaddress.ipaddress, e))
+            else:
+                self.debug("Failed to wget from VM=%s http server on public_ip=%s: %s" % (vm.name, public_ip.ipaddress.ipaddress, e))
+
+    def setup_webserver(self, vm):
+        # Start httpd service on VM first
+        sshClient = vm.get_ssh_client()
+        # Test to see if we are on a tiny linux box (using busybox)
+        res = "apache"
+        try:
+            res = str(sshClient.execute("busybox")).lower()
+        except Exception as e:
+            self.debug("no busybox {e}")
+        if "httpd" in res:
+            self.debug(f"using busybox httpd on VM {vm.name}/{vm.id}")
+            self.setup_busybox(sshClient)
+        else:
+            self.debug(f"using apache httpd on VM {vm.name}/{vm.id}")
+            self.setup_apache(sshClient)
+
+    def setup_busybox(self, sshClient):
+        """ Create a dummy test.html file and fire up the busybox web server """
+        sshClient.execute('echo test > test.html')
+        sshClient.execute("/usr/sbin/httpd")
+        self.debug("Setup webserver using busybox")
+
+    def setup_apache(self, sshClient):
+        sshClient.execute("iptables -F")
+        sshClient.execute('echo test > /var/www/html/test.html')
+        sshClient.execute("service httpd start")
+        time.sleep(5)
+        ssh_response = str(sshClient.execute("service httpd status")).lower()
+        self.debug("httpd service status is: %s" % ssh_response)
+        if "httpd: unrecognized service" in ssh_response or "inactive" in ssh_response:
+            ssh_res = sshClient.execute("yum install httpd -y")
+            if "Complete!" not in ssh_res:
+                raise Exception("Failed to install http server")
+            sshClient.execute("service httpd start")
+            time.sleep(5)
+            ssh_response = str(sshClient.execute("service httpd status")).lower()
+        if not "running" in ssh_response:
+            raise Exception("Failed to start httpd service")
+        self.debug("Setup webserver using apache")
+
+    def check_ssh_into_vm(self, vm, public_ip, testnegative=False):
+        self.debug(f"Checking if we can SSH into VM={vm.name} on public_ip={public_ip.ipaddress.ipaddress}")
+        try:
+            vm.get_ssh_client(ipaddress=public_ip.ipaddress.ipaddress)
+            if not testnegative:
+                self.debug(f"SSH into VM={vm.name} on public_ip={public_ip.ipaddress.ipaddress} was successful")
+            else:
+                self.fail(f"SSH into VM={vm.name} on public_ip={public_ip.ipaddress.ipaddress} succeeded, but should have been rejected")
+        except:
+            if not testnegative:
+                self.fail(f"Failed to SSH into VM {vm.name} with ip {public_ip.ipaddress.ipaddress}")
+            else:
+                self.debug(f"Failed to SSH into VM {vm.name} with {public_ip.ipaddress.ipaddress}, as expected")
+
+    def create_natrule_for_services(self, vm, public_ip, network, services=None):
+        self.debug(f"Creating NAT rule in network for vm {vm.name} with public IP {public_ip.ipaddress.ipaddress}")
+        if not services:
+            services = self.services["natrule"]
+        nat_rule = NATRule.create(self.apiclient,
+                                  vm,
+                                  services,
+                                  ipaddressid=public_ip.ipaddress.id,
+                                  openfirewall=False,
+                                  networkid=network.id,
+                                  vpcid=self.vpc.id
+                                  )
+        self.cleanup.append(nat_rule)
+        self.debug("Adding NetworkACL rules to make NAT rule accessible")
+        nwacl_nat = NetworkACL.create(self.apiclient,
+                                      networkid=network.id,
+                                      services=services,
+                                      traffictype='Ingress'
+                                      )
+        self.cleanup.append(nwacl_nat)
+        self.debug(f'nwacl_nat={nwacl_nat.__dict__}')
+        return nat_rule
